+++ lib/dalli/protocol/meta.rb
@@ -21,6 +21,14 @@ module Dalli
 
       private
 
+      def record_get_metrics(attributes, result, options)
+        value = result.is_a?(Array) ? result.first : result
+        miss = cache_nils?(options) ? value.equal?(::Dalli::NOT_FOUND) : value.nil?
+        attributes['value_bytesize'] = miss || value.nil? ? 0 : value.bytesize
+        attributes['hit_count'] = miss ? 0 : 1
+        attributes['miss_count'] = miss ? 1 : 0
+      end
+
       # * only supports single server
       # * only supports set at the moment
       # * doesn't support cas at the moment
@@ -205,7 +213,6 @@ module Dalli
       # rubocop:disable Metrics/AbcSize
       # rubocop:disable Metrics/CyclomaticComplexity
       # rubocop:disable Metrics/PerceivedComplexity
-      # rubocop:disable Metrics/MethodLength
       def get(key, options = nil)
         encoded_key, base64 = KeyRegularizer.encode(key)
         meta_options = meta_flag_options(options)
@@ -227,19 +234,13 @@ module Dalli
                    else
                      response_processor.meta_get_with_value(cache_nils: cache_nils?(options))
                    end
-          unless attributes.frozen?
-            value = result.is_a?(Array) ? result.first : result
-            attributes['value_bytesize'] = value.nil? ? 0 : value.bytesize
-            attributes['hit_count'] = value.nil? ? 0 : 1
-            attributes['miss_count'] = value.nil? ? 1 : 0
-          end
+          record_get_metrics(attributes, result, options) unless attributes.frozen?
           result
         end
       end
       # rubocop:enable Metrics/AbcSize
       # rubocop:enable Metrics/CyclomaticComplexity
       # rubocop:enable Metrics/PerceivedComplexity
-      # rubocop:enable Metrics/MethodLength
 
       def quiet_get_request(key, req_options = nil)
         encoded_key, base64 = KeyRegularizer.encode(key)
@@ -273,7 +274,6 @@ module Dalli
         end
       end
 
-      # rubocop:disable Metrics/AbcSize
       def gat(key, ttl, options = nil)
         ttl = TtlSanitizer.sanitize(ttl)
         encoded_key, base64 = KeyRegularizer.encode(key)
@@ -290,16 +290,10 @@ module Dalli
                    else
                      response_processor.meta_get_with_value(cache_nils: cache_nils?(options))
                    end
-          unless attributes.frozen?
-            value = result.is_a?(Array) ? result.first : result
-            attributes['value_bytesize'] = value.nil? ? 0 : value.bytesize
-            attributes['hit_count'] = value.nil? ? 0 : 1
-            attributes['miss_count'] = value.nil? ? 1 : 0
-          end
+          record_get_metrics(attributes, result, options) unless attributes.frozen?
           result
         end
       end
-      # rubocop:enable Metrics/AbcSize
 
       def touch(key, ttl)
         ttl = TtlSanitizer.sanitize(ttl)